+++ react/components/Modal/BottomBar.js
@@ -4,15 +4,20 @@ import PropTypes from 'prop-types'
 import styles from './modal.css'
 
 const BottomBar = props => {
-  const { children, showTopShadow, responsiveFullScreen } = props
+  const { children, showTopShadow, responsiveFullScreen, showBorder } = props
 
   return (
     <div
       className={`
-        flex justify-content bt b--muted-4 flex-row-reverse
-        ${responsiveFullScreen ? 'ph7 pv5 ph8-ns pv6-ns' : 'ph8 pv6'}
+        flex justify-content flex-row-reverse 
+        ${showBorder ? 'bt b--muted-4 ' : ''}
+        ${
+          responsiveFullScreen
+            ? 'ph7 pv5 ph8-ns pv6-ns '
+            : 'ph6 ph8-ns pv5 pv6-ns '
+        }
         ${styles.shadowTransition}
-        ${showTopShadow ? 'shadow-4' : ''}
+        ${showTopShadow ? 'shadow-4 ' : ''}
       `}>
       {children}
     </div>
@@ -23,6 +28,7 @@ BottomBar.propTypes = {
   showTopShadow: PropTypes.bool,
   children: PropTypes.node,
   responsiveFullScreen: PropTypes.bool,
+  showBorder: PropTypes.bool,
 }
 
 BottomBar.defaultProps = {